Fixed null pointer exception in RandomColorGenerator. Adjusted UI and separated the Schedule Filter mechanism from the Scheduler Display mechanism. Also fixed a bug in SemesterSelectionPanel that prevented the system from retrieving a new semester from the server.
git-svn-id: https://rocs.googlecode.com/svn/trunk@108 06a21a18-f190-11dd-aa46-671127399dce

+++ src/java/edu/rpi/rocs/client/ui/semesterselect/SemesterSelectionPanel.java
@@ -148,7 +148,7 @@ public class SemesterSelectionPanel extends VerticalPanel {
 				int i = semesterList.getSelectedIndex();
 				Integer id = Integer.valueOf(semesterList.getValue(i));
 				for(SemesterDescription desc : allSemesters) {
-					if(desc.getSemesterId() == id) {
+					if(desc.getSemesterId().intValue() == id.intValue()) {
 						selectedSemester = desc;
 						selectedSemesterDidChange();
 						break;
